ingest/requirements.txt
yfinance>=0.2.40
supabase>=2.7.0
python-dotenv>=1.0.0
pandas>=2.0.0
requests>=2.31.0
